fantasy/server/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
Pillow
httpx
